pyproject.toml
[project]
name = "fitness-agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai[google-genai]>=1.6.0",
    "crewai-tools>=1.6.0",
    "fastapi>=0.122.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
